+++ wemake_python_styleguide/visitors/ast/keywords.py
@@ -2,7 +2,7 @@
 
 import ast
 from collections import defaultdict
-from typing import ClassVar, Dict, List, Tuple, Type, Union
+from typing import ClassVar, Dict, List, Optional, Tuple, Type, Union
 
 from typing_extensions import final
 
@@ -21,6 +21,7 @@ from wemake_python_styleguide.violations.best_practices import (
     WrongKeywordViolation,
 )
 from wemake_python_styleguide.violations.consistency import (
+    ConsecutiveYieldsViolation,
     InconsistentReturnVariableViolation,
     InconsistentReturnViolation,
     InconsistentYieldViolation,
@@ -215,6 +216,10 @@ class WrongContextManagerVisitor(BaseNodeVisitor):
 
 
 @final
+@alias('visit_any_function', (
+    'visit_FunctionDef',
+    'visit_AsyncFunctionDef',
+))
 class GeneratorKeywordsVisitor(BaseNodeVisitor):
     """Checks how generators are defined and used."""
 
@@ -228,6 +233,22 @@ class GeneratorKeywordsVisitor(BaseNodeVisitor):
         ast.GeneratorExp,
     )
 
+    def __init__(self, *args, **kwargs) -> None:
+        """Here we store the information about ``yield`` locations."""
+        super().__init__(*args, **kwargs)
+        self._yield_locations: Dict[int, ast.Expr] = {}
+
+    def visit_any_function(self, node: AnyFunctionDef) -> None:
+        """
+        We use this visitor method to check for consecutive ``yield`` nodes.
+
+        Raises:
+            ConsecutiveYieldsViolation
+
+        """
+        self._check_consecutive_yields(node)
+        self.generic_visit(node)
+
     def visit_YieldFrom(self, node: ast.YieldFrom) -> None:
         """
         Visits `yield from` nodes.
@@ -240,6 +261,11 @@ class GeneratorKeywordsVisitor(BaseNodeVisitor):
         self._check_yield_from_empty(node)
         self.generic_visit(node)
 
+    def _check_consecutive_yields(self, node: AnyFunctionDef) -> None:
+        for sub in ast.walk(node):
+            if isinstance(sub, ast.Expr) and isinstance(sub.value, ast.Yield):
+                self._yield_locations[sub.value.lineno] = sub
+
     def _check_yield_from_type(self, node: ast.YieldFrom) -> None:
         if not isinstance(node.value, self._allowed_nodes):
             self.add_violation(IncorrectYieldFromTargetViolation(node))
@@ -249,9 +275,24 @@ class GeneratorKeywordsVisitor(BaseNodeVisitor):
             if not node.value.elts:
                 self.add_violation(IncorrectYieldFromTargetViolation(node))
 
+    def _post_visit(self) -> None:
+        previous_line: Optional[int] = None
+        previous_parent: Optional[ast.AST] = None
+
+        for line, node in self._yield_locations.items():
+            parent = get_parent(node)
+
+            if previous_line is not None:
+                if line - 1 == previous_line and previous_parent == parent:
+                    self.add_violation(ConsecutiveYieldsViolation(node.value))
+                    break
+
+            previous_line = line
+            previous_parent = parent
+
 
 @final
-@alias('visit_return_variable', (
+@alias('visit_any_function', (
     'visit_AsyncFunctionDef',
     'visit_FunctionDef',
 ))
@@ -266,7 +307,7 @@ class ConsistentReturningVariableVisitor(BaseNodeVisitor):
         ast.Name,
     )
 
-    def visit_return_variable(self, node: AnyFunctionDef) -> None:
+    def visit_any_function(self, node: AnyFunctionDef) -> None:
         """
         Helper to get all ``return`` variables in a function at once.